feat: 홈페이지(카테고리 선택X) ui 구현
