--- src/components/home/RatedProduct.tsx
@@ -1,60 +1,0 @@
-import ProductCard from "@/components/common/productcard/ProductCard";
-
-const tempRatedData = [
-  {
-    id: 1,
-    name: "다이슨 슈퍼소닉 블루",
-    image: "/images/supersonic.svg",
-    reviewCount: 129,
-    favoriteCount: 34,
-    rating: 4.7,
-  },
-  {
-    id: 2,
-    name: "Apple Watch 7",
-    image: "/images/appleWatch.svg",
-    reviewCount: 4961,
-    favoriteCount: 162,
-    rating: 4.9,
-  },
-  {
-    id: 3,
-    name: "헤라 블랙쿠션",
-    image: "/images/heraBlack.svg",
-    reviewCount: 432,
-    favoriteCount: 57,
-    rating: 4.9,
-  },
-  {
-    id: 4,
-    name: "우스티드 울 폴로 셔츠",
-    image: "/images/poloShirt.svg",
-    reviewCount: 318,
-    favoriteCount: 44,
-    rating: 4.7,
-  },
-  {
-    id: 5,
-    name: "돌화분",
-    image: "/images/flowerpot.svg",
-    reviewCount: 274,
-    favoriteCount: 35,
-    rating: 4.9,
-  },
-  {
-    id: 6,
-    name: "아디다스 퍼피렛 코어 블랙",
-    image: "/images/adidas.svg",
-    reviewCount: 738,
-    favoriteCount: 94,
-    rating: 4.9,
-  },
-]
-
-export default function RatedProduct() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
