+++ src/components/home/SortDropdown.tsx
@@ -1,25 +1,24 @@
 import Dropdown from "@/components/common/dropdown/Dropdown";
-import ProductCard from "@/components/common/productcard/ProductCard";
 
 type Item = {
   id: number;
   name: string;
 };
 
-const DropdownData: Item[] = [
+const SortData: Item[] = [
   { id: 1, name: "최신순" },
   { id: 2, name: "별점 높은순" },
   { id: 3, name: "좋아요순" }
 ];
 
-export default function CategorySelected () {
+export default function SortDropdown() {
   return (
     <div>
       <div className="flex flex-row">
         <div>
           전자기기의 모든 상품
         </div>
-        <Dropdown items={DropdownData} onSelect={(item) => console.log(item)}>
+        <Dropdown items={SortData} onSelect={(item) => console.log(item)}>
           <Dropdown.Button placeholder="최신순" variant={"small"} />
           <Dropdown.List />
         </Dropdown>
